Criando o Token

+++ Controllers/CategoryController.cs
@@ -1,5 +1,7 @@
 ﻿using Blog.Data;
+using Blog.Extensions;
 using Blog.Models;
+using Blog.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -11,8 +13,16 @@ namespace Blog.Controllers
         [HttpGet("v1/categories")]
         public async Task<IActionResult> GetAsync([FromServices] BlogDataContext context)
         {
-            var categories = await context.Categories.ToListAsync();
-            return Ok(categories);
+            try
+            {
+                var categories = await context.Categories.ToListAsync();
+
+                return Ok(new ResultViewModel<List<Category>>(categories));
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(500, value: new ResultViewModel<List<Category>>(errors: "05X04 - Falha interna do servidor!"));
+            }
         }
 
         [HttpGet("v1/categories/{id:int}")]
@@ -21,44 +31,62 @@ namespace Blog.Controllers
             try
             {
                 var category = await context.Categories.FirstOrDefaultAsync(x => x.Id == id);
-                return Ok(category);
+
+                if (category == null)
+                {
+                    return NotFound(new ResultViewModel<Category>(errors: "Conteúdo não encontrado!"));
+                }
+
+                return Ok(new ResultViewModel<Category>(category));
             }
-            catch (DbUpdateException e)
+            catch
             {
-                return StatusCode(500, "05XE5 - Não foi possível obter a categoria!");
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, "05XE6 - Falha interna do servidor!");
+                return StatusCode(500, value: new ResultViewModel<Category>(errors: "05XE6 - Falha interna do servidor!"));
             }
         }
 
         [HttpPost("v1/categories")]
-        public async Task<IActionResult> PostAsync([FromBody] Category model, [FromServices] BlogDataContext context)
+        public async Task<IActionResult> PostAsync([FromBody] EditorCategoryViewModel model, [FromServices] BlogDataContext context)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(new ResultViewModel<Category>(ModelState.GetErrors()));
+            }
+
             try
             {
-                await context.Categories.AddAsync(model);
+                var category = new Category
+                {
+                    Id = 0,
+                    Name = model.Name,
+                    Slug = model.Slug.ToLower()
+                };
+                await context.Categories.AddAsync(category);
                 await context.SaveChangesAsync();
 
-                return Created($"v1/categories/{model.Id}", model);
+                return Created($"v1/categories/{category.Id}", new ResultViewModel<Category>(category));
             }
             catch (DbUpdateException e)
             {
-                return StatusCode(500, "05XE7 - Não foi possível Incluir a categoria!");
+                return StatusCode(500, new ResultViewModel<Category> ("05XE7 - Não foi possível Incluir a categoria!"));
             }
-            catch (Exception ex)
+            catch
             {
-                return StatusCode(500, "05XE8 - Falha interna do servidor!");
+                return StatusCode(500, new ResultViewModel<Category>("05XE8 - Falha interna do servidor!"));
             }
         }
 
         [HttpPut("v1/categories/{id:int}")]
-        public async Task<IActionResult> PutAsync([FromRoute] int id, [FromBody] Category model, [FromServices] BlogDataContext context)
+        public async Task<IActionResult> PutAsync([FromRoute] int id, [FromBody] EditorCategoryViewModel model, [FromServices] BlogDataContext context)
         {
             try
             {
                 var category = await context.Categories.FirstOrDefaultAsync(x => x.Id == id);
+
+                if (category == null)
+                {
+                    return NotFound(new ResultViewModel<Category>("Conteúdo não encontrado!"));
+                }
 
                 category.Name = model.Name;
                 category.Slug = model.Slug;
@@ -66,15 +94,15 @@ namespace Blog.Controllers
                 context.Categories.Update(category);
                 await context.SaveChangesAsync();
 
-                return Ok();
+                return Ok(new ResultViewModel<Category>(category));
             }
             catch (DbUpdateException e)
             {
-                return StatusCode(500, "05XE9 - Não foi possível incluir a categoria!");
+                return StatusCode(500, new ResultViewModel<Category>("05XE9 - Não foi possível alterar a categoria!"));
             }
             catch (Exception ex)
             {
-                return StatusCode(500, "05XE10 - Falha interna do servidor!");
+                return StatusCode(500, new ResultViewModel<Category>("05XE10 - Falha interna do servidor!"));
             }
         }
 
@@ -85,18 +113,23 @@ namespace Blog.Controllers
             {
                 var category = await context.Categories.FirstOrDefaultAsync(x => x.Id == id);
 
+                if (category == null)
+                {
+                    return NotFound(new ResultViewModel<Category>("Conteúdo não encontrado!"));
+                }
+
                 context.Categories.Remove(category);
                 await context.SaveChangesAsync();
 
-                return Ok(category);
+                return Ok(new ResultViewModel<Category>(category));
             }
             catch (DbUpdateException e)
             {
-                return StatusCode(500, "05XE11 - Não foi possível excluir a categoria!");
+                return StatusCode(500, new ResultViewModel<Category>("05XE11 - Não foi possível excluir a categoria!"));
             }
             catch (Exception ex)
             {
-                return StatusCode(500, "05XE12 - Falha interna do servidor!");
+                return StatusCode(500, new ResultViewModel<Category>("05XE12 - Falha interna do servidor!"));
             }
         }
     }
